Add more subtle NonText color in vim-vwilight
The more subtle scheme is better for making newlines less obtrusive, but
since NonText also applies to the non-text parts of the buffer display,
including everything after the end of the file, it effects the overall
feel of the view. As a result, leaving both versions in the file for
easier experimentation.